refact: сдел рефакторинг renderLetter и renderMessage
