--- script/clear.js
@@ -1,3 +1,0 @@
-export function clear(objHtmlElements) {
-  objHtmlElements.content.firstElementChild.remove(); 
-}
