--- script/newindex.js
@@ -1,9 +1,0 @@
-export {objHtmlElements, initialSenders, wrapperMailbox} from './data.js';
-export {getRandInt} from './getRandomInt.js';
-export {Letter, addLetter, initialLetters, setLetter, getLetters, openLetter, setIdLetters} from './newLetter.js';
-export {clear} from './clear.js';
-export {addMailSection, SetEventListenerMailbox} from './mailsection.js';
-
-
-
-
