+++ script/render/renderMenu.js
@@ -1,0 +1,5 @@
+export function renderMenu(objHtmlElements) {
+  const menuTemplate = document.querySelector(objHtmlElements.templaitMenu).content;
+  const menuElement = menuTemplate.cloneNode(true);
+  return menuElement;
+}
